Modified the way in which Handlers are specified and loaded.
It is not possible to add additional handlers by simply adding jars containing the handlers to the
sdata.war deployment. In each jar, if there is a META-INF/sdata.config file, it will be loaded and the
properties contained within the jar will be added to the list of handlers.

This approach avoids needing to put all the handlers into 1 jar and avoids having to have all the configuration
in a single place. The only downside is that the sdata.war need to have the extra jars added to it as a deployment.

The format of the properties is the same as in the web.xml, namely
1.classname=org.sakaiproject.sdata.tool.JCRDumper
1.baseurl=/d
2.classname=org.sakaiproject.sdata.tool.json.JsonCHSHandler
2.baseurl=/c
2.function.m=org.sakaiproject.sdata.tool.functions.CHSNodeMetadata
2.function.cf=org.sakaiproject.sdata.tool.functions.CHSCreateFolder
2.function.hr=org.sakaiproject.sdata.tool.functions.CHSHideReleaseFunction
2.function.mv=org.sakaiproject.sdata.tool.functions.CHSMoveFunction
2.function.pr=org.sakaiproject.sdata.tool.functions.CHSPropertiesFunction
2.function.pm=org.sakaiproject.sdata.tool.functions.CHSPermissionsFunction
2.function.t=org.sakaiproject.sdata.tool.functions.CHSTaggingFunction
2.locationbase=
2.referencebase=/content
2.basepath=/

Note, there is no more training ; on each line and this file is proccessed as a java.util.Properties file.

Each numbered key defines a set of properties.
The classname is the name of the class of the handler
The baseurl is the url which maps to the handler.
Other properties are passed to the handler to enable it to configure its
own properties.


git-svn-id: https://source.sakaiproject.org/contrib/tfd/trunk/sdata@50853 fdecad78-55fc-0310-b1b2-d7d25cf747c9